[no ci] WIP, moving aarch64 generic timer functionality to its own crate
