+++ kernel/generic_timer_aarch64/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+authors = [
+    "Nathan Royer <nathan.royer.pro@gmail.com>",
+    "Kevin Boos <kevinaboos@gmail.com",
+]
+name = "generic_timer_aarch64"
+description = "Support for aarch64's generic timer and system counter"
+version = "0.1.0"
+edition = "2021"
+
+
+[target.'cfg(target_arch = "aarch64")'.dependencies]
+memory_structs = { path = "../memory_structs" }
+derive_more = "0.99.0"
+
+[features]
+default = [ "qemu_virt" ]
+qemu_virt = []
